refactor: splitting keyring-snap-client into keyring-snap-internal-client + move rpc structs to the public keyring-api
