+++ packages/keyring-internal-api/tsconfig.json
@@ -3,6 +3,11 @@
   "compilerOptions": {
     "baseUrl": "./"
   },
-  "include": ["./src"],
+  "references": [
+    {
+      "path": "../keyring-api/tsconfig.build.json"
+    }
+  ],
+  "include": ["./src", "../keyring-api/src/api.ts"],
   "exclude": ["./dist/**/*"]
 }
