+++ packages/keyring-snap-bridge/src/SnapKeyring.ts
@@ -30,7 +30,7 @@ import type {
   EthUserOperationPatch,
 } from '@metamask/keyring-api';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
-import { KeyringSnapControllerClient } from '@metamask/keyring-snap-client';
+import { KeyringSnapControllerClient } from '@metamask/keyring-snap-internal-client';
 import type { SnapController } from '@metamask/snaps-controllers';
 import type { SnapId } from '@metamask/snaps-sdk';
 import type { Snap } from '@metamask/snaps-utils';
