+++ packages/keyring-snap-bridge/tsconfig.build.json
@@ -13,7 +13,7 @@
   "references": [
     { "path": "../keyring-api/tsconfig.build.json" },
     { "path": "../keyring-snap-sdk/tsconfig.build.json" },
-    { "path": "../keyring-snap-client/tsconfig.build.json" },
+    { "path": "../keyring-snap-internal-client/tsconfig.build.json" },
     { "path": "../keyring-internal-api/tsconfig.build.json" }
   ],
   "include": ["./src/**/*.ts"],
