+++ packages/keyring-snap-client/src/index.ts
@@ -1,3 +1,2 @@
 export * from './KeyringClient';
-export * from './KeyringSnapControllerClient';
 export * from './KeyringSnapRpcClient';
