+++ packages/keyring-snap-client/tsconfig.build.json
@@ -14,9 +14,6 @@
     },
     {
       "path": "../keyring-snap-sdk/tsconfig.build.json"
-    },
-    {
-      "path": "../keyring-api/tsconfig.build.json"
     }
   ],
   "include": ["./src/**/*.ts"],
