+++ packages/keyring-snap-internal-client/README.md
@@ -1,0 +1,15 @@
+# MetaMask Keyring Snap Clients
+
+Clients that can be used to communicate with a Snap Account using various providers.
+
+## Installation
+
+`yarn add @metamask/keyring-snap-internal-client`
+
+or
+
+`npm install @metamask/keyring-snap-internal-client`
+
+## Contributing
+
+This package is part of a monorepo. Instructions for contributing can be found in the [monorepo README](https://github.com/MetaMask/accounts#readme).
