+++ packages/keyring-snap-internal-client/src/index.ts
@@ -1,0 +1,1 @@
+export * from './KeyringSnapControllerClient';
