+++ packages/keyring-snap-internal-client/tsconfig.build.json
@@ -11,6 +11,12 @@
   "references": [
     {
       "path": "../keyring-api/tsconfig.build.json"
+    },
+    {
+      "path": "../keyring-snap-client/tsconfig.build.json"
+    },
+    {
+      "path": "../keyring-snap-sdk/tsconfig.build.json"
     }
   ],
   "include": ["./src/**/*.ts"],
