+++ packages/keyring-snap-internal-client/typedoc.json
@@ -1,0 +1,6 @@
+{
+  "entryPoints": ["./src/index.ts"],
+  "excludePrivate": true,
+  "hideGenerator": true,
+  "out": "docs"
+}
