+++ packages/keyring-snap-sdk/src/rpc-handler.test.ts
@@ -1,6 +1,5 @@
-import type { Keyring } from '@metamask/keyring-api';
 import { KeyringRpcMethod, isKeyringRpcMethod } from '@metamask/keyring-api';
-import type { GetAccountBalancesRequest } from '@metamask/keyring-internal-api';
+import type { Keyring, GetAccountBalancesRequest } from '@metamask/keyring-api';
 
 import type { JsonRpcRequest } from './JsonRpcRequest';
 import { handleKeyringRequest } from './rpc-handler';
