+++ tsconfig.json
@@ -9,6 +9,7 @@
     { "path": "./packages/keyring-eth-hd" },
     { "path": "./packages/keyring-snap-bridge" },
     { "path": "./packages/keyring-snap-client" },
+    { "path": "./packages/keyring-snap-internal-client" },
     { "path": "./packages/keyring-snap-sdk" },
     { "path": "./packages/keyring-utils" }
   ],
